+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## Upcoming
 ## Added
 - `TryGet` extension for `IDictionary`
+- `TryFirst` extension for `IEnumerable`
+- `TryLast` extension for `IEnumerable`
+- `TryElementAt` extension for `IEnumerable`
 ## 2.0.1
 ### Fixed
 - `FromNullable` method did not work with some types, like `string` but now it  does